# src/api/openapi.yaml
openapi: 3.1.0
info:
  title: Connection Manager API
  description: Connection Manager API
  version: 1.6.0
# Permission names are prefixed with this service identity: mcm.<operationId>
...
servers:
  - url: /api
...
paths:
...
  /dfsps/states-status:
    get:
      tags:
      - dfsp-config
      - ping
      - status
      summary: Get connectivity status for all DFSPs
      description: |
        Returns connectivity status including ping status and state status information for all DFSPs.
        Combines data from dfsps table (pingStatus, lastUpdatedPingStatusAt) with dfsp_states_status table.
      operationId: handleGetDfspsStatesStatus
      # Rows are per DFSP, so the caller sees only the DFSPs in its scope
      x-authz:
        scopedBy: [dfsps]
      security:
      - session: []
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DfspsConnectivityStatusResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /dfsps/endpoints/unprocessed:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Returns the unprocessed endpoint items
      operationId: getUnprocessedEndpointItems
      # Items belong to a DFSP, so the caller sees only the DFSPs in its scope
      x-authz:
        scopedBy: [dfsps]

      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
...
  /dfsps/{dfspId}/endpoints:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Returns all DFSP endpoints
      operationId: getDFSPEndpoints
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/unprocessed:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Returns the unprocessed dfsp items
      operationId: getUnprocessedDfspItems
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/ingress:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Get the DFSP Ingress configuration
      operationId: getDFSPIngress
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDFSPIngressRes'
      security:
      - session: []
      - machineToken: []
    post:
      tags:
      - DFSP-Endpoint-configuration
      summary: Adds DFSP Ingress endpoint configuration
      description: Adds the entry with state NEW
      operationId: createDFSPIngress
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      requestBody:
        description: DFSP ingress IP entry
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDFSPIngressReq'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDFSPIngressRes'
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/ingress/ips:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Get the DFSP Ingress IPs
      operationId: getDFSPIngressIps
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EndPointIp'
      security:
      - session: []
      - machineToken: []
    post:
      tags:
      - DFSP-Endpoint-configuration
      summary: Adds a new IP entry to the DFSP Ingress endpoint
      description: Adds the entry with state NEW
      operationId: createDFSPIngressIp
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      requestBody:
        description: DFSP ingress IP entry
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputIPValue'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointIp'
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/ingress/ips/{epId}:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Get an endpoint entry
      operationId: getDFSPIngressIpEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
    put:
      tags:
      - DFSP-Endpoint-configuration
      summary: Update an endpoint entry
      operationId: updateDFSPIngressIpEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      requestBody:
        description: endpoint entry
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputIPValue'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
    delete:
      tags:
      - DFSP-Endpoint-configuration
      summary: Delete an endpoint entry
      operationId: deleteDFSPIngressIpEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      responses:
        204:
          description: successful operation
          content: {}
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/ingress/urls:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Get the DFSP Ingress URLs
      operationId: getDFSPIngressUrls
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EndPointURL'
      security:
      - session: []
      - machineToken: []
    post:
      tags:
      - DFSP-Endpoint-configuration
      summary: Adds a DFSP Ingress URL
      description: Adds a DFSP ingress URL with state NEW
      operationId: createDFSPIngressUrl
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      requestBody:
        description: DFSP ingress URL
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputURLValue'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointURL'
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/ingress/urls/{epId}:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Get an endpoint entry
      operationId: getDFSPIngressUrlEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
    put:
      tags:
      - DFSP-Endpoint-configuration
      summary: Update an endpoint entry
      operationId: updateDFSPIngressUrlEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      requestBody:
        description: endpoint entry
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputURLValue'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
    delete:
      tags:
      - DFSP-Endpoint-configuration
      summary: Delete an endpoint entry
      operationId: deleteDFSPIngressUrlEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      responses:
        204:
          description: successful operation
          content: {}
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/egress:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Get the DFSP Egress configuration
      operationId: getDFSPEgress
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDFSPEgressRes'
      security:
      - session: []
      - machineToken: []
    post:
      tags:
      - DFSP-Endpoint-configuration
      summary: Adds DFSP Egress endpoint configuration
      description: Adds the entry with state NEW
      operationId: createDFSPEgress
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      requestBody:
        description: DFSP egress IP entry
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDFSPEgressReq'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDFSPEgressRes'
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/egress/ips:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Get the DFSP Egress IPs
      operationId: getDFSPEgressIps
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EndPointIp'
      security:
      - session: []
      - machineToken: []
    post:
      tags:
      - DFSP-Endpoint-configuration
      summary: Adds a new IP entry to the DFSP Egress endpoint
      description: Adds the entry with state NEW
      operationId: createDFSPEgressIp
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      requestBody:
        description: DFSP egress IP entry
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputIPValue'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointIp'
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/egress/ips/{epId}:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Get an endpoint entry
      operationId: getDFSPEgressIpEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
    put:
      tags:
      - DFSP-Endpoint-configuration
      summary: Update an endpoint entry
      operationId: updateDFSPEgressIpEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      requestBody:
        description: endpoint entry
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputIPValue'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
    delete:
      tags:
      - DFSP-Endpoint-configuration
      summary: Delete an endpoint entry
      operationId: deleteDFSPEgressIpEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      responses:
        204:
          description: successful operation
          content: {}
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/{epId}:
    get:
      tags:
      - DFSP-Endpoint-configuration
      summary: Get an endpoint entry
      operationId: getDFSPEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
    put:
      tags:
      - DFSP-Endpoint-configuration
      summary: Update an endpoint entry
      operationId: updateDFSPEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      requestBody:
        description: endpoint entry
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndPointAny'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
    delete:
      tags:
      - DFSP-Endpoint-configuration
      summary: Delete an endpoint entry
      operationId: deleteDFSPEndpoint
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: Endpoint entry id
        required: true
        schema:
          type: string
      responses:
        204:
          description: successful operation
          content: {}
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/endpoints/{epId}/confirmation:
    post:
      tags:
      - DFSP-Endpoint-configuration
      summary: Updates the endpoint as confirmed
      operationId: confirmEndpointItem
      parameters:
      - name: dfspId
        in: path
        description: DFSP id
        required: true
        schema:
          type: string
      - name: epId
        in: path
        description: ID of the endpoint item
        required: true
        schema:
          type: number
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndPointAny'
      security:
      - session: []
      - machineToken: []
...
  /dfsps/{dfspId}/status:
    get:
      summary: Retrieves dfsp status
      description: Retrieves dfsp status by phase and step completion
      operationId: getDfspStatus
      parameters:
      - name: dfspId
        in: path
        description: ID of the dfsp to get satus
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DfspStatus'
      security:
      - session: []
      - machineToken: []
  /dfsps/{dfspId}/states:
    post:
      tags:
        - states
      summary: Upload status of DFSP states
      description: Upload status of DFSP states
      operationId: uploadDfspStatesStatus
      security:
      - session: []
      - machineToken: []
      parameters:
        - name: dfspId
          in: path
          description: DFSP id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DfspStatesStatusPayload'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        type:
          type: string
      examples:
      - code: 0
        message: message
        type: type
...
    URLEntry:
      type: object
      additionalProperties: false
      properties:
        url:
          type: string
          description: URL
...
    EndPointIp:
      allOf:
      - $ref: '#/components/schemas/BaseEndpoint'
      - type: object
        properties:
          value:
            $ref: '#/components/schemas/IPEntry'
    EndPointURL:
      allOf:
      - $ref: '#/components/schemas/BaseEndpoint'
      - type: object
        properties:
          value:
            $ref: '#/components/schemas/URLEntry'
    EndPointAny:
      allOf:
      - $ref: '#/components/schemas/BaseEndpoint'
      - type: object
        properties:
          value:
            type: object
            description: It's *either* {url} OR {address and ports}.
            allOf:
            - $ref: '#/components/schemas/IPEntry'
            - $ref: '#/components/schemas/URLEntry'
    InputIPValue:
      type: object
      required:
      - value
      properties:
        value:
          $ref: '#/components/schemas/IPEntry'
...
    CreateDFSPIngressReq:
      $ref: '#/components/schemas/DFSPIngress'
      examples:
      - url: http://test.local
    GetDFSPIngressRes:
      allOf:
      - $ref: '#/components/schemas/BaseEndpointv2'
      - $ref: '#/components/schemas/DFSPIngress'
      examples:
      - {
          "id": 230,
          "dfspId": "test1",
          "state": "NOT_STARTED",
          "type": null,
          "createdAt": "2022-03-25T09:10:19.000Z",
          "createdBy": null,
          "url": "http://test.local"
        }
    CreateDFSPIngressRes:
      allOf:
      - $ref: '#/components/schemas/BaseEndpointv2'
      - $ref: '#/components/schemas/DFSPIngress'
      examples:
      - {
          "id": 230,
          "dfspId": "test1",
          "state": "NOT_STARTED",
          "type": null,
          "createdAt": "2022-03-25T09:10:19.000Z",
          "createdBy": null,
          "url": "http://test.local"
        }
    CreateDFSPEgressReq:
      $ref: '#/components/schemas/DFSPEgress'
      examples:
      - {
          "ipList": [
            {
              "description": "Notification Callback Egress IP & Ports",
              "address": "163.10.24.28/30",
              "ports": [
                "80",
                "8000-8080"
              ]
            }
          ]
        }
...
    InputURLValue:
      type: object
      required:
        - value
      properties:
        value:
          $ref: '#/components/schemas/URLEntry'
...
    DfspsConnectivityStatusResponse:
      type: object
      properties:
        dfsps:
          type: array
          items:
            $ref: '#/components/schemas/DfspConnectivityStatusItem'
      required:
        - dfsps
      additionalProperties: false
...
  securitySchemes:
    # A human's browser session, issued by Ory Kratos.
    session:
      type: apiKey
      in: cookie
      name: ory_kratos_session
    # A machine client's access token, issued by Ory Hydra.
    machineToken:
      type: http
      scheme: bearer
      bearerFormat: JWT